Split the PythonMIPBuilder class into a separate file
