+++ python/ommx-python-mip-adapter/ommx_python_mip_adapter/__init__.py
@@ -1,5 +1,6 @@
+from .ommx_to_python_mip import PythonMIPBuilder, instance_to_model
+
 from ommx_python_mip_adapter.adapter import (
-    instance_to_model,
     model_to_instance,
     model_to_solution,
 )
@@ -8,4 +9,5 @@ __all__ = [
     "instance_to_model",
     "model_to_instance",
     "model_to_solution",
+    "PythonMIPBuilder",
 ]
